Extend Temperature Conversion Table
The current limits of the temperature table are not sufficient. Given that the temperature sensor of the AC unit is reading higher than 30C, or 86F, the temperature value will not be reported correctly. Simply extending this range solves this problem. This is specifically to address the temperature sensor in the unit, not its climate-addressable range.