fix: Fahrenheit conversion inconsistencies (#73)
* If value not in generated table, pick a close value that is in the table.

* Extend Temperature Conversion Table

The current limits of the temperature table are not sufficient. Given that the temperature sensor of the AC unit is reading higher than 30C, or 86F, the temperature value will not be reported correctly. Simply extending this range solves this problem. This is specifically to address the temperature sensor in the unit, not its climate-addressable range.

* Add some tests for temRec mismatche

* Adjust temRec being sent to device on test.

* Widen the range of the calculated temperature table.

* Raise temperature to above new limit (100F->150F) for Fahrenheit.

* Update test_device.py

---------

Co-authored-by: Clifford Roche <1007595+cmroche@users.noreply.github.com>